refactor: genesis keypair files generation is done by genesis tools
